+++ lib/hooks/useFormSubmissionData.tsx
@@ -9,7 +9,7 @@ import { slugify } from "@lib/client/clientHelpers";
 import { getStartLabels } from "@lib/utils/form-builder/i18nHelpers";
 import { type HTMLProps } from "@lib/saveAndResume/types";
 
-import { copyObjectExcludingFileContent } from "./useResponseCache";
+import { copyObjectExcludingFileContent } from "../fileExtractor";
 
 import { getValuesWithMatchedIds, getVisibleGroupsBasedOnValuesRecursive } from "@gcforms/core";