+++ lib/hooks/useValueChanged.tsx
@@ -13,7 +13,6 @@ export const useFormValuesChanged = () => {
       // skip for test env
       return;
     }
-
     updateValues({ formValues: values as Record<string, string> });
   }, [updateValues, values, setFieldValue]);
 };